DOMA-1564 made fixes after self review
